requirements.txt
phidata
python-dotenv
yfinanace
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai